Add BrowserOnly for CI
